# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/accounts/overview": {
      "get": {
        "operationId": "AccountsController_getCommonAccountInfo",
        "summary": "Get account info, us market cap, unrealized P&L [23603200- OLTSAPI0025]",
        "parameters": [
          {
            "name": "SYSTEM_ID",
            "required": false,
            "in": "query",
            "description": "Order placement channel",
            "schema": {
              "example": "RU",
              "enum": [
                "AG",
                "SV",
                "CH",
                "AP",
                "MJ",
                "XT",
                "CW",
                "BT",
                "MM",
                "WM",
                "ES",
                "H5",
                "RU"
              ],
              "type": "string"
            }
          },
          {
            "name": "IP",
            "required": true,
            "in": "query",
            "description": "Customer IP",
            "schema": {
              "example": "00.00.00.00",
              "type": "string"
            }
          },
          {
            "name": "CUSTOMER_ID",
            "required": true,
            "in": "query",
            "description": "Customer account",
            "schema": {
              "example": "20000010854",
              "type": "string"
            }
          },
          {
            "name": "ID",
            "required": true,
            "in": "query",
            "description": "Customer ID",
            "schema": {
              "example": "L220289830",
              "type": "string"
            }
          },
          {
            "name": "MARKET",
            "required": false,
            "in": "query",
            "description": "Trading market",
            "schema": {
              "example": "",
              "enum": [
                "",
                "HK",
                "US",
                "SH",
                "JP",
                "TH",
                "KR",
                "GB",
                "FR",
                "DE"
              ],
              "type": "string"
            }
          },
          {
            "name": "PROD_TYPE",
            "required": false,
            "in": "query",
            "description": "Product Type",
            "schema": {
              "example": "EQTY",
              "enum": [
                "",
                "EQTY",
                "FUND",
                "DEBT"
              ],
              "type": "string"
            }
          },
          {
            "name": "SETTLE_CURRENCY",
            "required": false,
            "in": "query",
            "description": "Settlement currency",
            "schema": {
              "example": ""
            }
          },
          {
            "name": "TRADE_KIND",
            "required": true,
            "in": "query",
            "description": "Transaction type",
            "schema": {
              "example": 0,
              "enum": [
                0,
                1,
                2,
                4
              ],
              "type": "number"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "Market api token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "responseTime": {
                          "type": "string",
                          "example": "2024-01-01 23:59:59"
                        },
                        "data": {
                          "$ref": "#/components/schemas/GetAccountOverviewResponse"
                        },
                        "errors": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "accounts"
        ]
      }
    },
    "/accounts": {
      "get": {
        "operationId": "AccountsController_getAccounts",
        "summary": "Get list account infomation [25900000- OLTSAPI0024]",
        "parameters": [
          {
            "name": "SYSTEM_ID",
            "required": false,
            "in": "query",
            "description": "Order placement channel",
            "schema": {
              "example": "RU",
              "enum": [
                "AG",
                "SV",
                "CH",
                "AP",
                "MJ",
                "XT",
                "CW",
                "BT",
                "MM",
                "WM",
                "ES",
                "H5",
                "RU"
              ],
              "type": "string"
            }
          },
          {
            "name": "IP",
            "required": true,
            "in": "query",
            "description": "Customer IP",
            "schema": {
              "example": "00.00.00.00",
              "type": "string"
            }
          },
          {
            "name": "CUSTOMER_ID",
            "required": true,
            "in": "query",
            "description": "Customer account",
            "schema": {
              "example": "20000010854",
              "type": "string"
            }
          },
          {
            "name": "ID",
            "required": true,
            "in": "query",
            "description": "Customer ID",
            "schema": {
              "example": "L220289830",
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "Market api token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "responseTime": {
                          "type": "string",
                          "example": "2024-01-01 23:59:59"
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/GetAccountInfoResponse"
                          }
                        },
                        "errors": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "accounts"
        ]
      }
    },
    "/accounts/tab-records": {
      "get": {
        "operationId": "AccountsController_getMenu",
        "summary": "Get no. of records in menu account page 委託(orders) 成交(trades) [23603200- OLTSAPI0025, 21110004- OLTSAPI0026, 21120004- OLTSAPI0031]",
        "parameters": [
          {
            "name": "SYSTEM_ID",
            "required": false,
            "in": "query",
            "description": "Order placement channel",
            "schema": {
              "example": "RU",
              "enum": [
                "AG",
                "SV",
                "CH",
                "AP",
                "MJ",
                "XT",
                "CW",
                "BT",
                "MM",
                "WM",
                "ES",
                "H5",
                "RU"
              ],
              "type": "string"
            }
          },
          {
            "name": "IP",
            "required": true,
            "in": "query",
            "description": "Customer IP",
            "schema": {
              "example": "00.00.00.00",
              "type": "string"
            }
          },
          {
            "name": "CUSTOMER_ID",
            "required": true,
            "in": "query",
            "description": "Customer account",
            "schema": {
              "example": "20000010854",
              "type": "string"
            }
          },
          {
            "name": "ID",
            "required": true,
            "in": "query",
            "description": "Customer ID",
            "schema": {
              "example": "L220289830",
              "type": "string"
            }
          },
          {
            "name": "MARKET",
            "required": false,
            "in": "query",
            "description": "Trading market",
            "schema": {
              "example": "",
              "enum": [
                "",
                "HK",
                "US",
                "SH",
                "JP",
                "TH",
                "KR",
                "GB",
                "FR",
                "DE"
              ],
              "type": "string"
            }
          },
          {
            "name": "PROD_TYPE",
            "required": false,
            "in": "query",
            "description": "Product Type",
            "schema": {
              "example": "EQTY",
              "enum": [
                "",
                "EQTY",
                "FUND",
                "DEBT"
              ],
              "type": "string"
            }
          },
          {
            "name": "SETTLE_CURRENCY",
            "required": false,
            "in": "query",
            "description": "Settlement currency",
            "schema": {
              "example": ""
            }
          },
          {
            "name": "TRADE_KIND",
            "required": true,
            "in": "query",
            "description": "Transaction type",
            "schema": {
              "example": 0,
              "enum": [
                0,
                1,
                2,
                4
              ],
              "type": "number"
            }
          },
          {
            "name": "SYMBOL",
            "required": false,
            "in": "query",
            "description": "Symbol",
            "schema": {
              "example": "",
              "type": "string"
            }
          },
          {
            "name": "TRADE_DATES",
            "required": false,
            "in": "query",
            "description": "Trading date - from",
            "schema": {
              "example": "2021/10/06",
              "type": "string"
            }
          },
          {
            "name": "TRADE_DATEE",
            "required": false,
            "in": "query",
            "description": "Trading date - to",
            "schema": {
              "example": "2024/10/06",
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "Market api token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "responseTime": {
                          "type": "string",
                          "example": "2024-01-01 23:59:59"
                        },
                        "data": {
                          "$ref": "#/components/schemas/GetTabRecordResponse"
                        },
                        "errors": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "accounts"
        ]
      }
    },
    "/accounts/statement": {
      "get": {
        "operationId": "AccountsController_getStatement",
        "summary": "Get statement of account [22003200- OLTSAPI0027]",
        "parameters": [
          {
            "name": "SYSTEM_ID",
            "required": false,
            "in": "query",
            "description": "Order placement channel",
            "schema": {
              "example": "RU",
              "enum": [
                "AG",
                "SV",
                "CH",
                "AP",
                "MJ",
                "XT",
                "CW",
                "BT",
                "MM",
                "WM",
                "ES",
                "H5",
                "RU"
              ],
              "type": "string"
            }
          },
          {
            "name": "IP",
            "required": true,
            "in": "query",
            "description": "Customer IP",
            "schema": {
              "example": "00.00.00.00",
              "type": "string"
            }
          },
          {
            "name": "CUSTOMER_ID",
            "required": true,
            "in": "query",
            "description": "Customer account",
            "schema": {
              "example": "20000010854",
              "type": "string"
            }
          },
          {
            "name": "ID",
            "required": true,
            "in": "query",
            "description": "Customer ID",
            "schema": {
              "example": "L220289830",
              "type": "string"
            }
          },
          {
            "name": "SYMBOL",
            "required": false,
            "in": "query",
            "description": "Symbol",
            "schema": {
              "example": "",
              "type": "string"
            }
          },
          {
            "name": "TRADE_DATES",
            "required": true,
            "in": "query",
            "description": "Trading date - from",
            "schema": {
              "example": "2021/10/06",
              "type": "string"
            }
          },
          {
            "name": "TRADE_DATEE",
            "required": true,
            "in": "query",
            "description": "Trading date - to",
            "schema": {
              "example": "2024/10/06",
              "type": "string"
            }
          },
          {
            "name": "QUERY_TYPE",
            "required": true,
            "in": "query",
            "description": "Query Type",
            "schema": {
              "example": 0,
              "enum": [
                0,
                1,
                2,
                3
              ],
              "type": "number"
            }
          },
          {
            "name": "PROD_TYPE",
            "required": false,
            "in": "query",
            "description": "Product Type",
            "schema": {
              "example": "EQTY",
              "enum": [
                "",
                "EQTY",
                "FUND",
                "DEBT"
              ],
              "type": "string"
            }
          },
          {
            "name": "SRC_TYPE",
            "required": false,
            "in": "query",
            "description": "交易來源",
            "schema": {
              "example": "",
              "enum": [
                "",
                "ALP"
              ],
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "Market api token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "responseTime": {
                          "type": "string",
                          "example": "2024-01-01 23:59:59"
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/GetStatementResponse"
                          }
                        },
                        "errors": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "accounts"
        ]
      }
    },
    "/accounts/non-trading": {
      "get": {
        "operationId": "AccountsController_getNonTradingData",
        "summary": "Get non trading data [22003202- OLTSAPI0028]",
        "parameters": [
          {
            "name": "SYSTEM_ID",
            "required": false,
            "in": "query",
            "description": "Order placement channel",
            "schema": {
              "example": "RU",
              "enum": [
                "AG",
                "SV",
                "CH",
                "AP",
                "MJ",
                "XT",
                "CW",
                "BT",
                "MM",
                "WM",
                "ES",
                "H5",
                "RU"
              ],
              "type": "string"
            }
          },
          {
            "name": "IP",
            "required": true,
            "in": "query",
            "description": "Customer IP",
            "schema": {
              "example": "00.00.00.00",
              "type": "string"
            }
          },
          {
            "name": "CUSTOMER_ID",
            "required": true,
            "in": "query",
            "description": "Customer account",
            "schema": {
              "example": "20000010854",
              "type": "string"
            }
          },
          {
            "name": "ID",
            "required": true,
            "in": "query",
            "description": "Customer ID",
            "schema": {
              "example": "L220289830",
              "type": "string"
            }
          },
          {
            "name": "SYMBOL",
            "required": false,
            "in": "query",
            "description": "Symbol",
            "schema": {
              "example": "",
              "type": "string"
            }
          },
          {
            "name": "TRADE_DATES",
            "required": true,
            "in": "query",
            "description": "Trading date - from",
            "schema": {
              "example": "2021/10/06",
              "type": "string"
            }
          },
          {
            "name": "TRADE_DATEE",
            "required": true,
            "in": "query",
            "description": "Trading date - to",
            "schema": {
              "example": "2024/10/06",
              "type": "string"
            }
          },
          {
            "name": "SRC_TYPE",
            "required": false,
            "in": "query",
            "description": "交易來源",
            "schema": {
              "example": "",
              "enum": [
                "",
                "ALP"
              ],
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "Market api token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "responseTime": {
                          "type": "string",
                          "example": "2024-01-01 23:59:59"
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/GetNonTradingResponse"
                          }
                        },
                        "errors": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "accounts"
        ]
      }
    },
...
  },
  "info": {
    "title": "KGI Securities API spec",
    "description": "Internal APIs for KGI",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3057/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "apikey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "scheme": "apikey"
      }
    },
    "schemas": {
...
      "GetAccountOverviewResponse": {
        "type": "object",
        "properties": {
          "totalMarketCap": {
            "type": "number",
            "example": 44180.187548999995
          },
          "totalProfits": {
            "type": "number",
            "example": 1669.7283198500015
          },
          "totalReturnRate": {
            "type": "number",
            "example": 3.9048039163807404
          }
        },
        "required": [
          "totalMarketCap",
          "totalProfits",
          "totalReturnRate"
        ]
      },
      "GetAccountInfoResponse": {
        "type": "object",
        "properties": {
          "CUSTOMER_ID": {
            "type": "string",
            "example": "20001365441"
          },
          "CUSTOMER_NAME": {
            "type": "string",
            "example": "米奇"
          },
          "ADDRESS": {
            "type": "string",
            "example": "北市中山區明水路700號"
          },
          "AUTH_NAME": {
            "type": "string",
            "example": ""
          },
          "BRANCH": {
            "type": "string",
            "example": "桃園分公司 桃園成功"
          },
          "SALES": {
            "type": "string",
            "example": "林ＯＯ"
          },
          "G_SALES": {
            "type": "string",
            "example": "翁ＯＯ"
          },
          "BANK_NAME": {
            "type": "string",
            "example": "TSIB TPE"
          },
          "BANK_ACCOUNT": {
            "type": "string",
            "example": "006719102772"
          },
          "BANK_ACCOUNT2": {
            "type": "string",
            "example": "006756062778"
          },
          "BANK_NAME_TWD": {
            "type": "string",
            "example": "N/A"
          },
          "BANK_ACCOUNT_TWD": {
            "type": "string",
            "example": "20061100184612"
          },
          "OPENDATE": {
            "type": "string",
            "example": "2013/01/17"
          },
          "ESTATEMENT": {
            "type": "string",
            "example": "1test@kgi.com.tw"
          },
          "E_MAIL": {
            "type": "string",
            "example": ""
          },
          "CUSTOMER_TYPE": {
            "type": "number",
            "example": 1
          },
          "W8_FORM": {
            "type": "string",
            "example": "Y"
          },
          "ETRADE": {
            "type": "string",
            "example": "Y"
          },
          "ETRADE_PWDNO": {
            "type": "number",
            "example": 9999999999
          },
          "ETRADE_APPLYDATE": {
            "type": "string",
            "example": "2013/01/17"
          },
          "ETRADE_CANCELDATE": {
            "type": "string",
            "example": "1900/01/01"
          },
          "ETRADE_PWDDATE": {
            "type": "string",
            "example": "1900/01/01"
          },
          "BRANCH_CODE": {
            "type": "string",
            "example": "000"
          },
          "ID": {
            "type": "string",
            "example": "E121647731"
          },
          "W8W9_DATE": {
            "type": "string",
            "example": "2016/05/27"
          },
          "W9": {
            "type": "string",
            "example": "N"
          },
          "RISK_LEVEL": {
            "type": "string",
            "example": "穩健積極型"
          },
          "RISK_LEVEL_DATE": {
            "type": "string",
            "example": "2013/01/17"
          },
          "PI_STATUS": {
            "type": "string",
            "example": "A"
          },
          "HYDEBTFUND_RISK": {
            "type": "string",
            "example": "N"
          },
          "HYDEBTFUND_RISK_DATE": {
            "type": "string",
            "example": "1900/01/01"
          },
          "ACCT_TYPE": {
            "type": "string",
            "example": "D"
          },
          "PRE_ESTATEMENT": {
            "type": "string",
            "example": ""
          },
          "BANK_ACC_SIN": {
            "type": "string",
            "example": ""
          },
          "BANK_ID_SIN": {
            "type": "string",
            "example": ""
          },
          "SINGLE_APPLY": {
            "type": "string",
            "example": ""
          },
          "SINGLE_APPLY_DATE": {
            "type": "string",
            "example": ""
          },
          "KGI_ACCOUNT": {
            "type": "string",
            "example": ""
          },
          "DOCUMENT_STATUS": {
            "type": "string",
            "example": ""
          },
          "SIGNING_DATE": {
            "type": "string",
            "example": ""
          },
          "DOCUMENT_TYPE1_DATE": {
            "type": "string",
            "example": ""
          },
          "R_EQUITY_AGREEMENT": {
            "type": "string",
            "example": "N"
          },
          "R_EQUITY_AGREEMENT_DATE": {
            "type": "string",
            "example": ""
          },
          "PD_TRANSFER_AGREEMENT": {
            "type": "string",
            "example": "N"
          },
          "PD_SIGNING_DATE": {
            "type": "string",
            "example": ""
          },
          "QI_NATIONALITY": {
            "type": "string",
            "example": "1"
          },
          "SETTLE_TWD": {
            "type": "string",
            "example": "Y"
          },
          "TDCC_ACCOUNT": {
            "type": "string",
            "example": ""
          },
          "W_FLAG": {
            "type": "string",
            "example": ""
          }
        },
        "required": [
          "CUSTOMER_ID",
          "CUSTOMER_NAME",
          "ADDRESS",
          "AUTH_NAME",
          "BRANCH",
          "SALES",
          "G_SALES",
          "BANK_NAME",
          "BANK_ACCOUNT",
          "BANK_ACCOUNT2",
          "BANK_NAME_TWD",
          "BANK_ACCOUNT_TWD",
          "OPENDATE",
          "ESTATEMENT",
          "E_MAIL",
          "CUSTOMER_TYPE",
          "W8_FORM",
          "ETRADE",
          "ETRADE_PWDNO",
          "ETRADE_APPLYDATE",
          "ETRADE_CANCELDATE",
          "ETRADE_PWDDATE",
          "BRANCH_CODE",
          "ID",
          "W8W9_DATE",
          "W9",
          "RISK_LEVEL",
          "RISK_LEVEL_DATE",
          "PI_STATUS",
          "HYDEBTFUND_RISK",
          "HYDEBTFUND_RISK_DATE",
          "ACCT_TYPE",
          "PRE_ESTATEMENT",
          "BANK_ACC_SIN",
          "BANK_ID_SIN",
          "SINGLE_APPLY",
          "SINGLE_APPLY_DATE",
          "KGI_ACCOUNT",
          "DOCUMENT_STATUS",
          "SIGNING_DATE",
          "DOCUMENT_TYPE1_DATE",
          "R_EQUITY_AGREEMENT",
          "R_EQUITY_AGREEMENT_DATE",
          "PD_TRANSFER_AGREEMENT",
          "PD_SIGNING_DATE",
          "QI_NATIONALITY",
          "SETTLE_TWD",
          "TDCC_ACCOUNT",
          "W_FLAG"
        ]
      },
      "GetTabRecordResponse": {
        "type": "object",
        "properties": {
          "inventories": {
            "type": "number",
            "example": 3
          },
          "orders": {
            "type": "number",
            "example": 3
          },
          "trades": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "inventories",
          "orders",
          "trades"
        ]
      },
...
      "GetNonTradingResponse": {
        "type": "object",
        "properties": {
          "CUSTOMER_ID": {
            "type": "string",
            "example": "20001365441"
          },
          "CUSTOMER_NAME": {
            "type": "string",
            "example": "米奇"
          },
          "TRADE_DATE": {
            "type": "string",
            "example": "2021/10/10"
          },
          "SYMBOL_NAME": {
            "type": "string",
            "example": "蘋果"
          },
          "TRADE_CURRENCY": {
            "type": "string",
            "example": "USD"
          },
          "QTY": {
            "type": "number",
            "example": 0.71275
          },
          "AMOUNT": {
            "type": "number",
            "example": 1005
          },
          "NOTES": {
            "type": "string",
            "example": "定期定額買入"
          },
          "SETTLE_DATE": {
            "type": "string",
            "example": "2021/10/12"
          },
          "TRANS_TYPE": {
            "type": "string",
            "example": "股票撥入"
          },
          "DESCRIPTION": {
            "type": "string",
            "example": " Transfer from a/c Name:KGI Securities Co., Ltd, trust account"
          },
          "QUERY_TYPE": {
            "type": "string",
            "example": "4"
          },
          "STOCK_CLASS": {
            "type": "string",
            "example": ""
          }
        },
        "required": [
          "CUSTOMER_ID",
          "CUSTOMER_NAME",
          "TRADE_DATE",
          "SYMBOL_NAME",
          "TRADE_CURRENCY",
          "QTY",
          "AMOUNT",
          "NOTES",
          "SETTLE_DATE",
          "TRANS_TYPE",
          "DESCRIPTION",
          "QUERY_TYPE",
          "STOCK_CLASS"
        ]
      },
...
    }
  }
}